+++ frontend/app/hooks/transaction/useUpdateTransaction.tsx
@@ -3,11 +3,7 @@
 import { useState } from "react";
 import { axios } from "@/app/lib";
 
-interface UpdateTransactionReq {
-    title: string;
-    notes: string;
-    price: number;
-}
+import type { UpdateTransactionReq } from "@/app/types/http/UserRequest";
 
 export default function useUpdateTransaction() {
     const [loading, setLoading] = useState(false);
@@ -18,7 +14,12 @@ export default function useUpdateTransaction() {
         setError(null);
 
         try {
-            await axios.put(`/transactions/update/${id}`, data);
+            await axios.put(`/transactions/update/${id}`, {
+                title: data.title,
+                notes: data.notes,
+                price: Number(data.price),
+                reverse: data.reverse
+            });
             return true;
         } catch (err: any) {
             setError(err.message || "Failed to update transaction.");